requirements.txt
requests==2.28.2
transliterate==1.10.2
pytils==0.4.1
Pillow==9.4.0
